Order fn in core.utils and move some of them
